refactor: move "templates" pkg under "internal/pkg/template" pkg
