+++ internal/pkg/deploy/cloudformation/cloudformation.go
@@ -28,8 +28,6 @@ import (
 	"github.com/aws/copilot-cli/internal/pkg/stream"
 	"github.com/aws/copilot-cli/internal/pkg/term/log"
 	"github.com/aws/copilot-cli/internal/pkg/term/progress"
-	"github.com/aws/copilot-cli/templates"
-	"github.com/gobuffalo/packd"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -108,7 +106,6 @@ type CloudFormation struct {
 	ecsClient      ecsClient
 	regionalClient func(region string) cfnClient
 	appStackSet    stackSetClient
-	box            packd.Box
 	s3Client       s3Client
 	region         string
 }
@@ -126,7 +123,6 @@ func New(sess *session.Session) CloudFormation {
 			}))
 		},
 		appStackSet: stackset.New(sess),
-		box:         templates.Box(),
 		s3Client:    s3.New(sess),
 		region:      aws.StringValue(sess.Config.Region),
 	}
