+++ internal/pkg/template/templates/addons/aurora/cf.yml
