+++ internal/pkg/template/templates/addons/ddb/cf.yml
