+++ internal/pkg/template/templates/addons/s3/cf.yml
