+++ internal/pkg/template/templates/cicd/pipeline.yml
