+++ internal/pkg/template/templates/cicd/pipeline_cfn.yml
