+++ internal/pkg/template/templates/environment/partials/cfn-execution-role.yml
