+++ internal/pkg/template/templates/environment/partials/custom-resources-role.yml
