+++ internal/pkg/template/templates/environment/partials/custom-resources.yml
