+++ internal/pkg/template/templates/environment/partials/environment-manager-role.yml
