+++ internal/pkg/template/templates/environment/partials/lambdas.yml
