+++ internal/pkg/template/templates/environment/partials/nat-gateways.yml
