+++ internal/pkg/template/templates/environment/partials/vpc-resources.yml
