+++ internal/pkg/template/templates/workloads/jobs/scheduled-job/cf.yml
