+++ internal/pkg/template/templates/workloads/jobs/scheduled-job/manifest.yml
