+++ internal/pkg/template/templates/workloads/partials/cf/addons.yml
