+++ internal/pkg/template/templates/workloads/partials/cf/autoscaling.yml
