+++ internal/pkg/template/templates/workloads/partials/cf/env-controller.yml
