+++ internal/pkg/template/templates/workloads/partials/cf/envvars.yml
