+++ internal/pkg/template/templates/workloads/partials/cf/eventrule.yml
