+++ internal/pkg/template/templates/workloads/partials/cf/fargate-taskdef-base-properties.yml
