+++ internal/pkg/template/templates/workloads/partials/cf/image-overrides.yml
