+++ internal/pkg/template/templates/workloads/partials/cf/instancerole.yml
