+++ internal/pkg/template/templates/workloads/partials/cf/logconfig.yml
