+++ internal/pkg/template/templates/workloads/partials/cf/mount-points.yml
