+++ internal/pkg/template/templates/workloads/partials/cf/publish.yml
