+++ internal/pkg/template/templates/workloads/partials/cf/secrets.yml
