+++ internal/pkg/template/templates/workloads/partials/cf/service-base-properties.yml
