+++ internal/pkg/template/templates/workloads/partials/cf/servicediscovery.yml
