+++ internal/pkg/template/templates/workloads/partials/cf/state-machine-definition.json.yml
