+++ internal/pkg/template/templates/workloads/partials/cf/state-machine.yml
