+++ internal/pkg/template/templates/workloads/partials/cf/subscribe.yml
