+++ internal/pkg/template/templates/workloads/services/lb-web/manifest.yml
