+++ internal/pkg/template/templates/workloads/services/rd-web/manifest.yml
