+++ internal/pkg/template/templates/workloads/services/worker/cf.yml
