+++ internal/pkg/template/templates/workloads/services/worker/manifest.yml
